Add support for testing Dockerfile containers
Switch to a composite action since we can't scan docker containers from
our docker container. We need to just install mondoo onto the linux
runner. Also make sure we have rpm and then if it's a container scan
build the thing and scan it. If it's not a container do the same old
thing.

Signed-off-by: Tim Smith <tsmith84@gmail.com>